Move collider constraints into their own file
