+++ Assets/Code/Entities/Penguin/Components/PenguinColliderConstraints.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: e1a24bdf9ba8e2a448f61cb6606a45de
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
